Rename default_syscalls crate to seccomp_filters
This is a subcrate in the vmm which defined helper
functions for Firecracker to retrieve empty/default/custom
seccomp filters.

The old name was not really expressive.
