+++ src/firecracker/src/main.rs
@@ -17,8 +17,8 @@ use snapshot::Snapshot;
 use utils::arg_parser::{ArgParser, Argument};
 use utils::terminal::Terminal;
 use utils::validators::validate_instance_id;
-use vmm::default_syscalls::{get_custom_filters, get_default_filters, get_empty_filters};
 use vmm::resources::VmResources;
+use vmm::seccomp_filters::{get_custom_filters, get_default_filters, get_empty_filters};
 use vmm::signal_handler::{mask_handled_signals, SignalManager};
 use vmm::version_map::{FC_VERSION_TO_SNAP_VERSION, VERSION_MAP};
 use vmm::vmm_config::instance_info::InstanceInfo;
